Refactor Tests
